experiment/requirements.txt
boto3>=1.34.0
botocore>=1.34.0
python-dotenv>=1.0.1
Jinja2>=3.1.3
PyYAML>=6.0.2
rich>=13.7.1
